requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
Pillow==10.2.0
requests==2.31.0
firebase-admin==6.4.0
google-auth==2.27.0
dnspython==2.4.2
gunicorn==21.2.0